backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
gunicorn>=21.2.0
